+++ .env.example
@@ -2,3 +2,4 @@ REACT_APP_GLIFIC_API="http://glific.test:4000/api"
 REACT_APP_WEB_SOCKET="ws://glific.test:4000/socket"
 SENTRY_DSN=""
 REACT_APP_FLOW_EDITOR_API="http://glific.test:4000/flow-editor/"
+REACT_APP_FLOW_EDITOR_CONFIGURE="http://glific.test:3000/automation/configure"
