+++ src/common/constants.ts
@@ -8,6 +8,7 @@ export const REACT_APP_GLIFIC_AUTHENTICATION_API =
 export const USER_SESSION = process.env.REACT_APP_GLIFIC_API + '/v1/session';
 export const RESET_PASSWORD = process.env.REACT_APP_GLIFIC_API + '/v1/registration/reset-password';
 export const RENEW_TOKEN = USER_SESSION + '/renew';
+export const FLOW_EDITOR_CONFIGURE_LINK = process.env.REACT_APP_FLOW_EDITOR_CONFIGURE;
 
 // const enums
 // provider status against the contact
