+++ src/components/floweditor/FlowEditor.module.css
@@ -1,0 +1,19 @@
+.Button {
+  position: fixed !important;
+  bottom: 20px;
+  right: 15px;
+  z-index: 1000;
+}
+
+.Link {
+  text-decoration: none;
+}
+
+.HelpIcon {
+  position: fixed !important;
+  bottom: 25px;
+  right: 110px;
+  z-index: 1000;
+  width: 28px !important;
+  height: 28px;
+}
