--- src/components/floweditor/FlowEditorContainer/FlowEditorContainer.module.css
@@ -1,19 +1,0 @@
-.Button {
-  position: fixed !important;
-  bottom: 20px;
-  right: 15px;
-  z-index: 1000;
-}
-
-.Link {
-  text-decoration: none;
-}
-
-.HelpIcon {
-  position: fixed !important;
-  bottom: 25px;
-  right: 110px;
-  z-index: 1000;
-  width: 28px !important;
-  height: 28px;
-}
