+++ src/containers/Automation/AutomationList/AutomationList.tsx
@@ -32,7 +32,7 @@ const columnAttributes = {
 };
 const configureIcon = <ConfigureIcon></ConfigureIcon>;
 
-const additionalAction = { icon: configureIcon, parameter: 'id', link: '/automation/configure' };
+const additionalAction = { icon: configureIcon, parameter: 'uuid', link: '/automation/configure' };
 
 export const AutomationList: React.SFC<AutomationListProps> = (props) => (
   <List
